fix(team-detail): match user-detail edit/delete flow
Team detail's budget had two ways to edit (inline-on-blur + the
appbar Pencil edit-mode), which made the appbar Edit button feel
redundant. Drop the inline-editable input: in view mode the budget
now reads as plain currency (or "Not set" / a non-manager hint),
exactly like /users/[id]. Edit mode picks up the shared
formatBudgetInput live formatter and autoFocus too.

formatBudgetInput is now shared in lib/utils.ts so both pages stay
in sync.
